+++ crates/ruma-events/CHANGELOG.md
@@ -1,5 +1,13 @@
 # [unreleased]
 
+Improvements:
+
+ - Add support for encrypted stickers as sent by several bridges under the flag `compat-encrypted-stickers`
+
+Breaking changes:
+
+ - `StickerEventContent::url` was replaced by `StickerEventContent::source` which is a `StickerMediaSource`
+
 # 0.28.1
 
 Improvements: